+++ CHANGELOG.md
@@ -24,3 +24,4 @@ specification.
     This makes it different from `add`.
 - Action `add_prefix` allows to add prefixes to annotations and labels.
 - Util that intersects list of dicts by key value.
+- Methods `update_common` that update common annotations and common labels.